src/backend/requirements.txt
fastapi
uvicorn
boto3
python-dotenv
awscli
zipfile
shutil
re
fitz